Add helper functions to new package
Go does not allow importing code from a file marked as package main into
another file.
